Handle null foreign key values in belongsTo.
It's perfectly valid for a foreign key to be null. BookshelfJS just
returns null in such a case. But DataLoader's load method doesn't like
being passed null. An issue[1] was opened, but the DataLoader devs
closed it, suggesting instead to test for null before calling load. This
commit does exactly that, returning null instead of letting DataLoader
raise an exception when targetFK is null.

[1] https://github.com/facebook/dataloader/issues/98