Use query instead of where to add where clauses.
`model` can sometimes be a collection. `where` on a collection takes no
arguments and returns the whole collection for simple filtering, which makes it
an expensive noop in this case. `query`, on the other hand, always does the
right thing.

Includes tests both that we don't break querying fields without arguments and
that it works with arguments. Latter breaks without the fix, demonstrating the
original problem.